+++ src/app/students/components/students.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ConfirmationService } from 'primeng/api';
 import { Student } from '../domain-models/student.model';
 import { StudentService } from '../services/students.service';
 
@@ -9,18 +10,26 @@ import { StudentService } from '../services/students.service';
 })
 export class StudentsComponent implements OnInit {
   constructor(readonly studentSrv: StudentService,
+              private confirmationService: ConfirmationService,
               private _router: Router
   ) {}
 
   students: Array<Student> = [];
+  displayBasic: boolean = false;
 
   ngOnInit(): void {
     this.students = this.studentSrv.getStudents();    
   }
 
   deleteStudent(studentId: string): void {
-    this.studentSrv.deleteStudent(studentId);
-    this.students = this.studentSrv.getStudents();
+    this.confirmationService.confirm({
+      message: `Czy na pewno chcesz usunąć studenta o id: ${studentId}?`,
+      accept: () => {
+        this.studentSrv.deleteStudent(studentId);
+        this.students = this.studentSrv.getStudents();
+        this.displayBasic = true; 
+      }
+  }); 
   }
 
   editStudent(studentId: string) {
